+++ lib/appium_lib_core/common.rb
@@ -4,3 +4,4 @@ require_relative 'common/log'
 require_relative 'common/command'
 require_relative 'common/device'
 require_relative 'common/base'
+require_relative 'common/websocket'
